converting integration tests into unit tests
