+++ service/src/test/java/uk/nhs/adaptors/gp2gp/ehr/EhrResendControllerTest.java
@@ -7,41 +7,27 @@ import org.hl7.fhir.dstu3.model.OperationOutcome;
 import org.hl7.fhir.dstu3.model.UriType;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpStatus;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
 import uk.nhs.adaptors.gp2gp.common.service.RandomIdGeneratorService;
 import uk.nhs.adaptors.gp2gp.common.service.TimestampService;
 import uk.nhs.adaptors.gp2gp.common.task.TaskDispatcher;
 import uk.nhs.adaptors.gp2gp.ehr.model.EhrExtractStatus;
-import uk.nhs.adaptors.gp2gp.gpc.GetGpcStructuredTaskDefinition;
-import uk.nhs.adaptors.gp2gp.testcontainers.ActiveMQExtension;
-import uk.nhs.adaptors.gp2gp.testcontainers.MongoDBExtension;
 
 import java.time.Duration;
 import java.time.Instant;
-import java.time.temporal.ChronoUnit;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.hl7.fhir.dstu3.model.OperationOutcome.IssueType;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.mockito.Mockito.doReturn;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static uk.nhs.adaptors.gp2gp.ehr.EhrStatusConstants.INCUMBENT_NACK_CODE;
-import static uk.nhs.adaptors.gp2gp.ehr.EhrStatusConstants.INCUMBENT_NACK_DISPLAY;
 
-@SpringBootTest
-@DirtiesContext
-@ExtendWith({SpringExtension.class, MongoDBExtension.class, ActiveMQExtension.class, MockitoExtension.class})
+@ExtendWith(MockitoExtension.class)
 public class EhrResendControllerTest {
 
     public static final Instant NOW = Instant.parse("2024-01-01T10:00:00Z");
@@ -51,21 +37,23 @@ public class EhrResendControllerTest {
     public static final String NHS_NUMBER = "12345";
     private static final String TO_ASID_CODE = "test-to-asid";
     private static final String FROM_ASID_CODE = "test-from-asid";
+    public static final String INCUMBENT_NACK_CODE = "99";
+    public static final String INCUMBENT_NACK_DISPLAY = "Unexpected condition.";
 
-    @Autowired
+    @Mock
     private EhrExtractStatusRepository ehrExtractStatusRepository;
 
-    @Autowired
-    private EhrResendController ehrResendController;
+    @Mock
+    private TimestampService timestampService;
 
-    @MockBean
+    @Mock
     private RandomIdGeneratorService randomIdGeneratorService;
 
-    @MockBean
+    @Mock
     private TaskDispatcher taskDispatcher;
 
-    @MockBean
-    private TimestampService timestampService;
+    @InjectMocks
+    private EhrResendController ehrResendController;
 
     @Test
     public void When_AnEhrExtractHasNotFailedAndAnotherResendRequestArrives_Expect_FailedOperationOutcome() {
@@ -84,9 +72,12 @@ public class EhrResendControllerTest {
             .ehrRequest(EhrExtractStatus.EhrRequest.builder().nhsNumber(NHS_NUMBER).toAsid(TO_ASID_CODE).fromAsid(FROM_ASID_CODE).build())
             .build();
 
-        ehrExtractStatusRepository.save(IN_PROGRESS_EXTRACT_STATUS);
+        doReturn(Optional.of(IN_PROGRESS_EXTRACT_STATUS)).when(ehrExtractStatusRepository).findByConversationId(CONVERSATION_ID);
 
-        var operationOutcome = createOperationOutcome(IssueType.BUSINESSRULE, OperationOutcome.IssueSeverity.ERROR, details, diagnostics);
+        var operationOutcome = createOperationOutcome(OperationOutcome.IssueType.BUSINESSRULE,
+                                                      OperationOutcome.IssueSeverity.ERROR,
+                                                      details,
+                                                      diagnostics);
 
         var response = ehrResendController.scheduleEhrExtractResend(CONVERSATION_ID);
 
@@ -97,34 +88,6 @@ public class EhrResendControllerTest {
 
     @Test
     public void When_AnEhrExtractHasFailed_Expect_RespondsWith202() {
-
-        String ehrMessageRef = generateRandomUppercaseUUID();
-        var ehrExtractStatus = new EhrExtractStatus();
-
-        ehrExtractStatus.setConversationId(CONVERSATION_ID);
-        ehrExtractStatus.setEhrReceivedAcknowledgement(EhrExtractStatus.EhrReceivedAcknowledgement.builder()
-                                        .conversationClosed(FIVE_DAYS_AGO)
-                                        .errors(List.of(
-                                            EhrExtractStatus.EhrReceivedAcknowledgement.ErrorDetails.builder()
-                                                .code(INCUMBENT_NACK_CODE)
-                                                .display(INCUMBENT_NACK_DISPLAY)
-                                                .build()))
-                                        .messageRef(ehrMessageRef)
-                                        .received(FIVE_DAYS_AGO)
-                                        .rootId(generateRandomUppercaseUUID())
-                                        .build());
-        ehrExtractStatus.setEhrRequest(EhrExtractStatus.EhrRequest.builder().nhsNumber(NHS_NUMBER).build());
-
-        ehrExtractStatusRepository.save(ehrExtractStatus);
-
-        var response = ehrResendController.scheduleEhrExtractResend(CONVERSATION_ID);
-
-        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.ACCEPTED);
-        assertThat(response.getBody()).isNull();
-    }
-
-    @Test
-    public void When_AnEhrExtractHasFailed_Expect_GetGpcStructuredTaskScheduled() {
 
         String ehrMessageRef = generateRandomUppercaseUUID();
         var ehrExtractStatus = new EhrExtractStatus();
@@ -142,24 +105,13 @@ public class EhrResendControllerTest {
                                                            .rootId(generateRandomUppercaseUUID())
                                                            .build());
         ehrExtractStatus.setEhrRequest(EhrExtractStatus.EhrRequest.builder().nhsNumber(NHS_NUMBER).build());
-        ehrExtractStatus.setEhrExtractCorePending(EhrExtractStatus.EhrExtractCorePending.builder().build());
-        ehrExtractStatus.setEhrContinue(EhrExtractStatus.EhrContinue.builder().build());
 
-        ehrExtractStatusRepository.save(ehrExtractStatus);
-        Instant now = Instant.now().truncatedTo(ChronoUnit.MILLIS);
-        doReturn(now).when(timestampService).now();
+        doReturn(Optional.of(ehrExtractStatus)).when(ehrExtractStatusRepository).findByConversationId(CONVERSATION_ID);
 
-        ehrResendController.scheduleEhrExtractResend(CONVERSATION_ID);
+        var response = ehrResendController.scheduleEhrExtractResend(CONVERSATION_ID);
 
-        var updatedEhrExtractStatus = ehrExtractStatusRepository.findByConversationId(ehrExtractStatus.getConversationId());
-        var taskDefinition = GetGpcStructuredTaskDefinition.getGetGpcStructuredTaskDefinition(randomIdGeneratorService, ehrExtractStatus);
-
-        verify(taskDispatcher, times(1)).createTask(taskDefinition);
-        assertEquals(now, updatedEhrExtractStatus.get().getMessageTimestamp());
-        assertNull(updatedEhrExtractStatus.get().getEhrExtractCorePending());
-        assertNull(updatedEhrExtractStatus.get().getEhrContinue());
-        assertNull(updatedEhrExtractStatus.get().getAckPending());
-        assertNull(updatedEhrExtractStatus.get().getEhrReceivedAcknowledgement());
+        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.ACCEPTED);
+        assertNull(response.getBody());
     }
 
     @Test
@@ -172,7 +124,12 @@ public class EhrResendControllerTest {
         details.setCoding(List.of(codeableConceptCoding));
         var diagnostics = "Provide a conversationId that exists and retry the operation";
 
-        var operationOutcome = createOperationOutcome(IssueType.VALUE, OperationOutcome.IssueSeverity.ERROR, details, diagnostics);
+        var operationOutcome = createOperationOutcome(OperationOutcome.IssueType.VALUE,
+                                                      OperationOutcome.IssueSeverity.ERROR,
+                                                      details,
+                                                      diagnostics);
+        doReturn(Optional.empty()).when(ehrExtractStatusRepository).findByConversationId(CONVERSATION_ID);
+
 
         var response = ehrResendController.scheduleEhrExtractResend(CONVERSATION_ID);
 
@@ -198,6 +155,5 @@ public class EhrResendControllerTest {
             .setDiagnostics(diagnostics);
         return operationOutcome;
     }
-
 
 }
